File sync working, but no tests

+++ config.js
@@ -1,4 +1,5 @@
 'use strict';
+import fsExtra from 'fs-extra';
 import { promises as fs } from 'fs';
 import untildify from 'untildify';
 import path from 'path';
@@ -7,16 +8,14 @@ import os from 'os';
 const init = async () => {
   const configDirectory = path.join(os.homedir(), '.config/onedrive.js');
   const configFilename = path.join(configDirectory, 'config.json');
-  await fs.mkdir(configDirectory, { recursive: true });
-  try {
-    await fs.access(configFilename);
-  } catch (err) {
-    await fs.copyFile('./default.config.json', configFilename);
-  }
+  await fsExtra.ensureDir(configDirectory);
+  const configFileExists = await fsExtra.pathExists(configFilename);
+  if (!configFileExists) await fs.copyFile('./default.config.json', configFilename);
   const configFileBuffer = await fs.readFile(configFilename);
   const config = JSON.parse(configFileBuffer.toString());
   ['db', 'syncDirectory', 'cacheDirectory']
     .forEach(key => { config[key] = untildify(config[key]); });
+  await fsExtra.remove(config.cacheDirectory);
   return config;
 };
